% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RehospitalizationPredictionWithNote.R
\docType{package}
\name{RehospitalizationPredictionWithNote}
\alias{RehospitalizationPredictionWithNote}
\alias{RehospitalizationPredictionWithNote-package}
\title{RehospitalizationPredictionWithNote}
\description{
RehospitalizationPredictionWithNote
}
